feat: Add getSessionUser function for retrieving user session
The `getSessionUser` function has been added to the `utils.ts` file. This function retrieves the user session using the `getServerSession` method from the `next-auth/next` package and returns the user object and user ID. This change enables the retrieval of user information from the session, which can be used for various purposes such as displaying user-specific content or performing user-related actions.